Use an external version of STP
This patch modifies KLEE to use an external version of STP as specified
by the --with-stp configure option.  It includes documentation
referring users to the STP download location.

Apart from applying this patch, these two things also need to be done.
I did not include them in this patch for space reasons.
 - delete the entire contents of the stp directory
 - rerun the autotools using the autoconf/AutoRegen.sh script