Floating point rewriter improvements
Specifically, test whether c & ¬p (where c is some existing
constraint in the set and p is the new predicate) is infeasible by
constructing the expression and checking if it simplifies to false.
If so, p is valid under this set of constraints and we substitute
a valid constraint set.  This allows us to use some of the existing
floating point simplifications across constraints.